Reformate int and long

+++ src/main/java/ru/netology/javaqa/zadanie6/service/StatsService.java
@@ -3,7 +3,7 @@ package ru.netology.javaqa.zadanie6.service;
 public class StatsService {
 
     // Метод для расчета суммы всех продаж
-    public int calculateTotalSales(int[] sales) {
+    public long calculateTotalSales(int[] sales) {
         int total = 0;
         for (int sale : sales) {
             total += sale;
@@ -12,13 +12,13 @@ public class StatsService {
     }
 
     // Метод для расчета средней суммы продаж в месяц
-    public int calculateAverageSales(int[] sales) {
-        int total = calculateTotalSales(sales);
+    public long calculateAverageSales(int[] sales) {
+        long total = calculateTotalSales(sales);
         return (int) total / sales.length;
     }
 
     // Метод для подсчета количества месяцев, в которых продажи были ниже среднего
-    public int calculateBelowAverageSales(int[] sales) {
+    public long calculateBelowAverageSales(int[] sales) {
         double average = calculateTotalSales(sales);
         int count = 0;
         for (int sale : sales) {
@@ -30,7 +30,7 @@ public class StatsService {
     }
 
     // Метод для подсчета количества месяцев, в которых продажи были выше среднего
-    public int calculateLiveAboveAverage(int[] sales) {
+    public long calculateLiveAboveAverage(int[] sales) {
         double average = calculateTotalSales(sales);
         int count = 0;
         for (int sale : sales) {
@@ -42,7 +42,7 @@ public class StatsService {
     }
 
     // Метод с минимальными продажами среди просмотренных ранее
-    public int calculateMinSales(int[] sales) {
+    public long calculateMinSales(int[] sales) {
         int minMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[minMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
@@ -54,7 +54,7 @@ public class StatsService {
     }
 
     // Метод с максимальными продажами среди просмотренных ранее
-    public int calculateMaxSales(int[] sales) {
+    public long calculateMaxSales(int[] sales) {
         int maxMonth = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= sales[maxMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
