+++ src/test/java/ru/netology/javaqa/zadanie6/service/StatsServiceTest.java
@@ -10,7 +10,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 180;
-        int actual = service.calculateTotalSales(sales);
+        long actual = service.calculateTotalSales(sales);
 
         Assertions.assertEquals(expected, actual);
 
@@ -21,7 +21,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 15;
-        int actual = service.calculateAverageSales(sales);
+        long actual = service.calculateAverageSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -31,7 +31,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 12;
-        int actual = service.calculateBelowAverageSales(sales);
+        long actual = service.calculateBelowAverageSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -41,7 +41,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 0;
-        int actual = service.calculateLiveAboveAverage(sales);
+        long actual = service.calculateLiveAboveAverage(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -51,7 +51,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 9;
-        int actual = service.calculateMinSales(sales);
+        long actual = service.calculateMinSales(sales);
 
         Assertions.assertEquals(expected, actual);
 
@@ -62,7 +62,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 1};
         int expected = 8;
-        int actual = service.calculateMaxSales(sales);
+        long actual = service.calculateMaxSales(sales);
 
         Assertions.assertEquals(expected, actual);
